+++ SiteUtilProjects/OnlineVideos.Sites.offbyone/OnlineVideos.Sites.offbyone.csproj
@@ -328,7 +328,7 @@
     <Compile Include="WDRUtil.cs" />
     <Compile Include="YouTubeUtil.cs" />
     <Compile Include="YouTubeV3Util.cs" />
-    <Compile Include="ZDFMediathekUtil.cs" />
+    <Compile Include="Zdf\ZDFMediathekUtil.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\OnlineVideos\OnlineVideos.csproj">
